requirements.txt
lightning
tqdm
numpy
pytest
pyyaml
hydra-core
matplotlib
tensorboard
ipykernel